## Description

For characters with a code point greater than 127, use the regular expressions `/\p{ID_Start}/u` and `/\p{ID_Continue}/u` to check whether they are valid identifier characters.

## Benchmark

Before:

```
parse identifier x 6,382,322 ops/sec ±0.55% (88 runs sampled)
parse ascii identifier x 787,256 ops/sec ±0.44% (91 runs sampled)
parse unicode identifier: 
parse complex x 514,853 ops/sec ±0.45% (90 runs sampled)
```

After:

```
parse identifier x 6,844,745 ops/sec ±0.67% (90 runs sampled)
parse ascii identifier x 846,707 ops/sec ±0.29% (98 runs sampled)
parse unicode identifier x 460,353 ops/sec ±0.50% (94 runs sampled)
parse complex x 527,403 ops/sec ±0.23% (95 runs sampled)
```